pyproject.toml
[project]
name = "summarising-medical-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=6.0.4",
    "django-cors-headers>=4.9.0",
    "djangorestframework>=3.17.1",
    "djangorestframework-simplejwt>=5.5.1",
    "nltk>=3.9.4",
    "setuptools>=82.0.1",
    "wheel>=0.46.3",
]
